After clicking on Clone button
When we set everything and click on Clone button, if repository is large in size and due to any reason connection to repository losses, then cloning process starts from zero instead of resuming from last stage. This is very complicate if we clone and large repository and irritates if about 90% done the last process and connection fails. Please resolve this to save the developers time. Time is money.